# New hashmap implementation