First off, I'm sorry this comes shortly after ~~@squeek502 's~~ work on the robin hood hashmap. I have been working on and off on this hashmap implementation for quite some time with the goal to contribute it to the std.

However I think the two implementations are complementary and no work was wasted.
I tried to adhere as much as possible to the current API,

## Design
It's based on open addressing (all elements are stored in a single contiguous array) and linear probing (we resolve collisions by just trying the next slot in the array). Quite similar to the widely publicized google's swiss tables, but a lot simpler.

### 1. Fast
The goal is to have a hashmap that is as fast as possible for lookups (considered most important usecase) and insertion/removal (second most important).

#### Statistics
We assume that the hash function is of good quality, giving unbiased results and dispatching elements all over the available slots. This is an absolute prerequisite for hashmap implementations, and that is the case with Zig's standard hash function (though there are other candidates obviously).
The probability of an element being assigned to a slot is `1/number_of_slot`. This does not mean that we can have a bijection between a key and a slot, therefore we need to handle collisions.

Linear probing is an efficient (see next §) and _very simple_ way to deal with collisions. I quite like that it's very easy to understand, and has very predictable behavior for the CPU. However, when probing a collision chain to find the correct key, the algorithm needs to perform equality comparison on the keys. For simple types such as ints, that's not really a problem, but keys can be larger and much more complex (you don't want to do lots of string comparisons for example). To remedy this, the hashmap keeps 6 bits from each hash and stores it as a metadata by slot (along with its state : free, used, tombstone).

The mecanism to determine which ideal slot the key belongs to already uses `log2(number_of_slots)` bits from the hash ; since we always keep a power-of-two number of slots to do fast modulus by masking. Keys that belong in the same ideal slot will have their hashes with identical `log2(number_of_slots)` low bits. But since the hash function is assumed to give random results, the 6 high bits of their hashes are very likely to be different. Using this further helps to differenciate keys without resorting to equality comparison.

#### Spatial locality
Pieces of data that are accessed together benefit greatly from begin close together in memory as CPUs optimize for this usecase. By using only 8bits of metadata per element and storing all the metadata contiguously, when accessing one (while doing a lookup for example) we typically get the metadata of 7 other slots for free (assuming a 64bytes cache line).

Effectively this means that even if the hashmap is nearly full and has collisions that require probing, it is almost free since a probe chain of length 8 is already in cache.
Not having to access the slot array until we're certain to have found the correct slot also means that we don't waste memory bandwith with unused data. If the metadata were to be embedded within the slots, probing would make a bad use of cache as they would be espaced by the size of an element.

### 2. Memory efficient
8 bits of metadata is quite low, and it's hard to go further with comparable advantages.

The hashmap also holds only one allocation that contains both the metadata and slot array. This is hard to measure but I think this helps reduce pressure on the allocator and fragmentation.

### 3. Small footprint
Extra effort was spent in trying to keep the `struct` as small as possible : it's only 16 bytes (24 for the `Managed` variant).
The fields embedded in the struct are the ones needed most frequently to avoid unnecessary cache misses. Other field-candidates are stored in the allocation (mainly, the capacity).

To keep it at 16 bytes, the size of the hashmap is limited to 32 bits. This seems like a reasonnable choice to me.

## Comparison with status quo

### Pros
- in my benchmarks, almost always faster by a good margin
- uses less memory, only 8bits of overhead by element
- can achieve higher load factors for similar speed, for even lower memory use
### Cons
- iteration speed over elements is slower by an order of magnitude (you can't beat contiguous arrays !)
- absolutely no order guarantee (which can be nice, but is not a main goal of hashmaps in my opinion)
- modification invalidates live iterators

### Performance
I've taken the best of 3 runs from my benchmark, which is basically just a longer version of what's now used in gotta-go-fast. Ported from https://martin.ankerl.com/2019/04/01/hashmap-benchmarks-01-overview/

#### Master
```
iterate while adding 0.428s
iterate while removing 0.427s

insert 100M int 13.606s
clear 100M int 0.127s
reinsert 100M int 7.300s
remove 100M int 8.824s
reinsert 100M int 7.293s
deinit map 0.148s

5% distinct 9.440s
25% distinct 10.650s
50% distinct 11.089s
100% distinct 9.932s

0% success, ffffffff  6.095ns
0% success, ffffffff00000000  6.168ns
25% success, ffffffff  6.161ns
25% success, ffffffff00000000  6.190ns
50% success, ffffffff  6.186ns
50% success, ffffffff00000000  6.123ns
75% success, ffffffff  6.238ns
75% success, ffffffff00000000  6.097ns
100% success, ffffffff  6.608ns
100% success, ffffffff00000000  6.597ns
```

#### This PR
```
iterate while adding 6.430s
iterate while removing 7.732s

insert 100M int 9.489s
clear 100M int 0.005s
reinsert 100M int 5.474s
remove 100M int 3.331s
reinsert 100M int 6.953s
deinit map 0.452s

5% distinct 3.844s
25% distinct 5.801s
50% distinct 6.572s
100% distinct 8.181s

0% success, ffffffff  5.641ns
0% success, ffffffff00000000  5.986ns
25% success, ffffffff  5.543ns
25% success, ffffffff00000000  5.374ns
50% success, ffffffff  5.077ns
50% success, ffffffff00000000  4.923ns
75% success, ffffffff  4.614ns
75% success, ffffffff00000000  4.526ns
100% success, ffffffff  4.398ns
100% success, ffffffff00000000  4.413ns
```

### Right tool for the job
I think there's use for both implementations, and renamed the current one to `sliceable_hash_map` as it is its main advantage. That is only a proposal.