@squeek502 2020-08-08
Just to clarify, I didn't do any work on the new hash map implementation. That was all @andrewrk. I just made some graphs.