@Sahnvour 2020-08-09
There's still a few errors (for example I think the translate-c code depends on the ordering of the current hashmap somewhere), which I'll solve if there is interest in this.