@squeek502 2020-08-10
Ran [my basic insertion benchmark](https://github.com/squeek502/zig-hash-map-bench) out of curiosity and got some interesting results.

(note: this only goes up to 10 million insertions)

![5999-i32](https://user-images.githubusercontent.com/2389051/89817588-46f0f900-dafd-11ea-98a2-8f0b726758da.png)
![5999-strings](https://user-images.githubusercontent.com/2389051/89817590-48222600-dafd-11ea-97de-ae508267344e.png)

I keep coming back to this comment: https://github.com/ziglang/gotta-go-fast/issues/2#issuecomment-647170135

> because hash maps can have very different performance characteristics depending on the size of the map, a benchmark like 'add 1,000,000 elements, get 1,000,000 times, then remove all the elements one by one' would only represent one single point along the continuum of possible hash map sizes, and therefore it would fail to catch regressions in performance for other map sizes.