@Sahnvour 2020-08-10
That's true, and I totally agree that there are many (too many, in fact) dimensions on which we can compare hashmaps.
However that would be a belittlement of Martin's benchmarks which try to model some "real life" scenarios and have good considerations in my opinion. Granted, they are more interested in big hashmaps.

I ran your benchmark on my machine and got these (raw) results
![image](https://user-images.githubusercontent.com/1806407/89822210-1edeb580-db50-11ea-884e-dc48282a7183.png)
So I guess we can also add CPU arch, memory speed, OS, etc. to the mix.

For example, I think the low insertion counts (< 1000 ?) mostly measure allocation speed, rather than insertion.

I'm a bit puzzled by your results in this range, and can't quite yet wrap my head around the sawtooth pattern in my graph for this PR. Maybe the number of inserts are pathologically bad in every odd count (for example, _just_ needing a grow + rehash, thus increasing the mean insertion time).