@squeek502 2020-08-11
I'm a total beginner to this type of stuff so I'm more than willing to believe that I'm not taking the proper things into account. My graphs are very strange indeed.