@ifreund 2020-08-11
I haven't read through your implementation in detail, but from my knowledge of the swiss tables implementation they make good use of SIMD to speed up lookups. What advantages does your design have over such an approach? Or could that be a potential future improvement?