@Sahnvour 2020-08-12
@ifreund It is my understanding that google tries to achieve extremely high load factor in its hashmaps, because memory usage is literally money to them. I would have to watch their videos again so don't quote me on that but IIRC they go as high as 97.5% occupancy to reduce memory waste.

This means that even with a very good hash function that provides good distribution, they likely have long collision chains when their maps are full; because the set of possible input keys is almost always larger than the set of possible hash results. Flat hashmaps (as this one, or the stdandard one) performance relies heavily on reducing collision probability by using more memory than would be minimally needed to store all their elements. The more additional memory you reserve, the more there are slots that will be unused and thus break collision chains. This means than when looking for a slot (to either do a lookup or a modification) you can statistically stop traversing your data earlier.

When you allow very high load factors, the number of free slots is reduced to a bare minimum that you find acceptable (this really is just a tradeoff) and collision chains get longer and longer. This implies that when looking for a slot, you will likely have to probe for more already used slots before finding the good one. In both google and this implementation, probing is mainly done by looking at an array of 8bit metadata. When collision chains get longer, testing multiple metadata at once with SIMD becomes interesting.

However, doing probing by packs of say 16 metadata implies some complexification of the design and implementation (which are best explained in their conferences so I will not go on only by my memory of it :) that IMO is not needed with less extreme load factors to achieve good performance.
When I prototyped this implementation in C++, I was able to beat google's in pretty much all benchmarks from the suite I linked in my previous comment. Your mileage may vary though.

So to conclude, it might be an improvement but it also might not. 😄