@andrewrk 2020-08-19
> Just to clarify, I didn't do any work on the new hash map implementation. That was all @andrewrk. I just made some graphs.

they were pretty graphs tho :wink: 

> However I think the two implementations are complementary and no work was wasted.

I agree with this conclusion. However, let's try to make it really clear what the differences and usage patterns would be. I think it would be fair to say that the API of the master branch hash map is strictly more convenient. Having the ArrayList of entries available for direct access is pretty nice in terms of ergonomics, and the fact that order is preserved (and independent of the hash function) is a really nice property. So it would seem the benefits of this implementation are trading some of this API convenience for better resource usage (memory & CPU). This being the case, I'd like to make sure it actually is significantly better than status quo hash maps before adding it as another option. Based on some of the data here, it's not entirely clear, right? I think before merging this it would be worth it to understand the benefits in a more conclusive way, so that the doc comments in the std lib can confidently explain when to choose one over the other.

> sawtooth pattern

My guess is it has to do with the underlying memory allocator returning memory that it already has available vs requesting extra memory from the OS. E.g. imagine appending to an array list, and on the 9th element appended, it doubles its capacity.