@Sahnvour 2020-08-19
Well, to me the results are pretty clear in that this PR has better (up to 2.5-3x) performance on all operations except iteration. The results are focused on large hashmaps, but that is what we currently have to measure.
It would indeed be interesting to have some benchmarks targeting small hashmaps because that is a common usecase that can have a different performance profile.

Can we define the required level of insight you need to accept of reject this PR ?

Thinking again about the sawtooth, apart from the influence of the allocator, in my opinion this is normal and expected, but not necessarily representative. Since storage is not reserved for the hashmap in the benchmark, it has to grow when necessary. Depending on the number of elements inserted (x values), if a grow was just triggered for this number because it exceeded previous capacity, the `total time / number of inserts` (y values) will be higher than for higher numbers of inserts, until the hashmap has to grow again. The choice of x values just happen to illustrate that.
If we plotted the total time for the inserts instead, we would likely see some kind of plateaux.